+++ api/src/main/java/com/albert/commerce/product/command/application/CreatedProductResponse.java
@@ -1,12 +1,8 @@
-package com.albert.commerce.product.application;
+package com.albert.commerce.product.command.application;
 
 import com.albert.commerce.common.model.Money;
 import com.albert.commerce.product.command.domain.Product;
 import com.albert.commerce.product.command.domain.ProductId;
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateTimeDeserializer;
-import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -18,7 +14,7 @@ import org.springframework.hateoas.RepresentationModel;
 @Getter
 @Setter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ProductResponse extends RepresentationModel<ProductResponse> {
+public class CreatedProductResponse extends RepresentationModel<CreatedProductResponse> {
 
     private ProductId productId;
     private String productName;
@@ -26,31 +22,21 @@ public class ProductResponse extends RepresentationModel<ProductResponse> {
     private String description;
     private String brand;
     private String category;
-    @JsonDeserialize(using = LocalDateTimeDeserializer.class)
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyyMMddHHmmss")
-    private LocalDateTime createdTime;
-    @JsonDeserialize(using = LocalDateTimeDeserializer.class)
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyyMMddHHmmss")
-    private LocalDateTime updateTime;
 
     @Builder
-    private ProductResponse(ProductId productId, String productName, Money price,
+    private CreatedProductResponse(ProductId productId, String productName, Money price,
             String description,
-            String brand, String category, LocalDateTime createdTime, LocalDateTime updateTime) {
+            String brand, String category) {
         this.productId = productId;
         this.productName = productName;
         this.price = price;
         this.description = description;
         this.brand = brand;
         this.category = category;
-        this.createdTime = createdTime;
-        this.updateTime = updateTime;
     }
 
-    public static ProductResponse from(Product product) {
-        return ProductResponse.builder()
-                .createdTime(product.getCreatedTime())
-                .updateTime(product.getUpdateTime())
+    public static CreatedProductResponse from(Product product) {
+        return CreatedProductResponse.builder()
                 .productId(product.getProductId())
                 .productName(product.getProductName())
                 .brand(product.getBrand())
