+++ api/src/main/java/com/albert/commerce/store/command/application/SellerStoreResponse.java
@@ -1,4 +1,4 @@
-package com.albert.commerce.store.application;
+package com.albert.commerce.store.command.application;
 
 import com.albert.commerce.common.units.BusinessLinks;
 import com.albert.commerce.store.command.domain.Store;
