+++ api/src/main/java/com/albert/commerce/store/command/domain/StoreRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StoreRepository extends JpaRepository<Store, StoreId> {
 
-
+    boolean existsByStoreUserId(StoreUserId storeUserId);
 }
