+++ api/src/main/java/com/albert/commerce/store/query/StoreDao.java
@@ -3,12 +3,11 @@ package com.albert.commerce.store.query;
 import com.albert.commerce.store.command.domain.Store;
 import com.albert.commerce.store.command.domain.StoreId;
 import com.albert.commerce.store.command.domain.StoreUserId;
+import com.albert.commerce.store.infra.StoreDaoCustom;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StoreDao extends JpaRepository<Store, StoreId>, StoreDaoCustom {
-
-    boolean existsByStoreUserId(StoreUserId storeUserId);
 
     Optional<Store> findByStoreUserId(StoreUserId storeUserId);
 }
