+++ api/src/test/java/com/albert/commerce/store/application/SellerStoreServiceTest.java
@@ -8,10 +8,13 @@ import static org.mockito.Mockito.mock;
 
 import com.albert.commerce.common.model.SequenceGenerator;
 import com.albert.commerce.common.model.SequenceGeneratorImpl;
+import com.albert.commerce.store.command.application.NewStoreRequest;
+import com.albert.commerce.store.command.application.SellerStoreResponse;
+import com.albert.commerce.store.command.application.SellerStoreService;
+import com.albert.commerce.store.command.application.StoreAlreadyExistsException;
 import com.albert.commerce.store.command.domain.Store;
 import com.albert.commerce.store.command.domain.StoreId;
 import com.albert.commerce.store.command.domain.StoreRepository;
-import com.albert.commerce.store.query.StoreDao;
 import com.albert.commerce.user.command.domain.UserId;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -22,15 +25,13 @@ class SellerStoreServiceTest {
 
     private SellerStoreService sellerStoreService;
     private StoreRepository storeRepository;
-    private StoreDao storeDao;
     private SequenceGenerator sequenceGenerator;
 
     @BeforeEach
     void setUserService() {
         storeRepository = mock(StoreRepository.class);
-        storeDao = mock(StoreDao.class);
         sequenceGenerator = new SequenceGeneratorImpl();
-        sellerStoreService = new SellerStoreService(storeRepository, storeDao, sequenceGenerator);
+        sellerStoreService = new SellerStoreService(storeRepository, sequenceGenerator);
     }
 
     @DisplayName("User는 Store를 하나 만들 수 있습니다.")
@@ -55,7 +56,7 @@ class SellerStoreServiceTest {
         // given
         NewStoreRequest newStoreRequest = new NewStoreRequest("test");
         newStoreRequest.setUserId(UserId.from(sequenceGenerator.generate()));
-        given(storeDao.existsByStoreUserId(any())).willReturn(true);
+        given(storeRepository.existsByStoreUserId(any())).willReturn(true);
 
         // when,then
         assertThatThrownBy(() -> sellerStoreService.createStore(newStoreRequest)).isInstanceOf(
